requirements.txt
python-telegram-bot[webhooks]==21.6
requests==2.32.5